# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Apployd Control Plane API
  version: 0.1.0
  description: REST API for account, projects, deployments, billing, and observability.
servers:
  - url: http://localhost:4000/api/v1
paths:
...
  /deployments:
    get:
      tags: [Deployments]
      security: [{ bearerAuth: [] }]
      summary: List deployments for project
      parameters:
        - in: query
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deployment history }
    post:
      tags: [Deployments]
      security: [{ bearerAuth: [] }]
      summary: Trigger deployment
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
          description: Replay-safe key for deploy request deduplication (1 hour window).
      responses:
        '202': { description: Deployment queued }
        '200': { description: Idempotent replay response with existing deployment }
        '402': { description: Active subscription required }
        '409': { description: Resource pool constraint violation }
  /deployments/{deploymentId}:
    get:
      tags: [Deployments]
      security: [{ bearerAuth: [] }]
      summary: Get deployment status by id
      parameters:
        - in: path
          name: deploymentId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deployment status }
  /deployments/{deploymentId}/wake:
    post:
      tags: [Deployments]
      security: [{ bearerAuth: [] }]
      summary: Wake sleeping deployment container
      parameters:
        - in: path
          name: deploymentId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Wake in progress }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT